@@ -15,7 +15,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 0.0.8
+version: 0.1.0
 
 environment:
   sdk: ">=2.7.0 <3.0.0"
@@ -23,19 +23,22 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
+
   famedlysdk: # Matrix library
     git:
       url: https://gitlab.com/famedly/famedlysdk.git
-  redux: ^4.0.0
-  flutter_redux: 0.7.0
-  provider: ^4.3.2+3
+      ref: main
+
+  redux: ^5.0.0
+  flutter_redux: ^0.8.2
+  provider: ^5.0.0
   
   # web navigator for SSO
-  webview_flutter: ^1.0.7
-  url_launcher: ^5.7.10
+  webview_flutter: ^2.0.2
+  url_launcher: ^6.0.3
 
-  flutter_typeahead: ^1.9.1
-  cached_network_image: ^2.4.1
+  flutter_typeahead: ^3.1.1
+  
   
   # Post formating
   flutter_markdown:
@@ -45,16 +48,20 @@ dependencies:
   file_picker_cross:
 
   # time formating
-  timeago: ^2.0.29
+  timeago: ^3.0.2
 
   # storage
-  moor: 3.4.0
-  sqlite3_flutter_libs: ^0.3.0
+  moor: ^4.0.0
+  sqlite3_flutter_libs: ^0.4.1
 
-  sqflite: ^1.3.2+1 # Still used to obtain the database location
-  sqlite3: ^0.1.8
-  localstorage: ^3.0.3+6
-  flutter_secure_storage: ^3.3.5
+  sqflite: ^2.0.0+3 # Still used to obtain the database location
+  sqlite3: ^1.0.0
+  localstorage: ^4.0.0+1
+  flutter_secure_storage: ^4.1.0
+  
+  # cache 
+  flutter_cache_manager: ^3.0.1
+  cached_network_image: ^3.0.0
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
@@ -66,14 +73,18 @@ dependencies:
   # Flushbar
   flushbar: ^1.10.4
   #dialogs
-  adaptive_dialog: ^0.9.3
+  adaptive_dialog: ^0.10.0+5
+  logging: ^1.0.1
 
-  flutter_olm: ^1.0.1
+  flutter_olm: ^1.1.1
+  package_info_plus: ^1.0.0
   
 dev_dependencies:
-  moor_generator: 3.4.0
   flutter_test:
     sdk: flutter
+
+
+
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
