@@ -7,24 +7,47 @@ import 'package:minestrix/global/smatrix.dart';
 import 'package:minestrix/global/smatrixWidget.dart';
 import 'package:timeago/timeago.dart' as timeago;
 
-class RepliesVue extends StatelessWidget {
+class RepliesVue extends StatefulWidget {
   final Event event;
   final Set<Event> replies;
   final String regex = "(>(.*)\n)*\n";
-  const RepliesVue({Key key, @required this.event, @required this.replies})
+  final bool showEditBox;
+  RepliesVue(
+      {Key key,
+      @required this.event,
+      @required this.replies,
+      this.showEditBox = false})
       : super(key: key);
+
+  @override
+  _RepliesVueState createState() => _RepliesVueState();
+}
+
+class _RepliesVueState extends State<RepliesVue> {
+  bool showEditBox = null;
+
   @override
   Widget build(BuildContext context) {
+    if (showEditBox == null) showEditBox = widget.showEditBox;
+
     // get replies
     SClient sclient = Matrix.of(context).sclient;
-    int max = min(replies.length, 2);
+    int max = min(widget.replies.length, 2);
 
     return Container(
 //      decoration: BoxDecoration(color: Colors.grey),
       child: Column(
         crossAxisAlignment: CrossAxisAlignment.start,
         children: [
-          for (Event revent in replies.toList().sublist(0, max))
+          if (widget.showEditBox)
+            ReplyBox(
+                event: widget.event,
+                onMessageSend: () {
+                  setState(() {
+                    showEditBox = false;
+                  });
+                }),
+          for (Event revent in widget.replies.toList().sublist(0, max))
             Padding(
               padding: const EdgeInsets.symmetric(vertical: 2.0),
               child: Column(
@@ -73,8 +96,8 @@ class RepliesVue extends StatelessWidget {
                                         ],
                                       ),
                                       SizedBox(height: 5),
-                                      Text(revent.body
-                                          .replaceFirst(new RegExp(regex), "")),
+                                      Text(revent.body.replaceFirst(
+                                          new RegExp(widget.regex), "")),
                                     ],
                                   ),
                                 ),
@@ -91,12 +114,12 @@ class RepliesVue extends StatelessWidget {
                         event: revent,
                         replies: revent.aggregatedEvents(
                             sclient.srooms[revent.roomId].timeline,
-                            RelationshipTypes.Reply)),
+                            RelationshipTypes.reply)),
                   )
                 ],
               ),
             ),
-          if (replies.length > max)
+          if (widget.replies.length > max)
             Center(
                 child:
                     MaterialButton(child: Text("load more"), onPressed: () {}))
@@ -108,7 +131,10 @@ class RepliesVue extends StatelessWidget {
 
 class ReplyBox extends StatelessWidget {
   final Event event;
-  const ReplyBox({Key key, @required this.event}) : super(key: key);
+  final Function onMessageSend;
+
+  const ReplyBox({Key key, @required this.event, this.onMessageSend})
+      : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -139,6 +165,9 @@ class ReplyBox extends StatelessWidget {
               onPressed: () async {
                 await event.room.sendTextEvent(tc.text, inReplyTo: event);
                 tc.clear();
+
+                // send event
+                onMessageSend();
               })
         ],
       ),
