@@ -1,4 +1,4 @@
-package com.example.minestrix
+package fr.henri2h.minestrix
 
 import io.flutter.embedding.android.FlutterActivity
 
