@@ -5,21 +5,36 @@
 import 'dart:async';
 
 import 'package:famedlysdk/famedlysdk.dart';
+import 'package:logging/logging.dart';
+import 'package:minestrix/global/smatrix/Notifications.dart';
 import 'package:minestrix/global/smatrix/SMatrixRoom.dart';
 
 class SClient extends Client {
+  final log = Logger("SClient");
+
   static const String SMatrixRoomPrefix = "smatrix_";
   static const String SMatrixUserRoomPrefix = SMatrixRoomPrefix + "@";
+
   StreamSubscription onEventUpdate;
   StreamSubscription onRoomUpdateSub; // event subscription
+
   StreamController<String> onTimelineUpdate = StreamController.broadcast();
+
+  // This stream is used to notify the user that a new post has been added to the timeline
+  StreamController<String> onNewPostInTimeline = StreamController.broadcast();
+
+  StreamController<String> onSRoomsUpdate = StreamController.broadcast();
 
   Map<String, SMatrixRoom> srooms = Map<String, SMatrixRoom>();
 
   // room sub types
   Map<String, SMatrixRoom> get sgroups => Map.from(srooms)
     ..removeWhere((key, value) => value.roomType != SRoomType.Group);
+
   Map<String, SMatrixRoom> get sfriends => Map.from(srooms)
+    ..removeWhere((key, value) => value.roomType != SRoomType.UserRoom);
+
+  Map<String, SMatrixRoom> get following => Map.from(srooms)
     ..removeWhere((key, value) => value.roomType != SRoomType.UserRoom);
 
   Map<String, SMatrixRoom> sInvites =
@@ -29,8 +44,8 @@ class SClient extends Client {
   List<Event> stimeline = [];
 
   bool _firstSync = true;
-  // ignore: unused_field
-  Timer _timer; // timer used to sync all the conversations on the first run
+
+  Notifications notifications = Notifications();
 
   SClient(String clientName,
       {bool enableE2eeRecovery,
@@ -40,7 +55,7 @@ class SClient extends Client {
             verificationMethods: verificationMethods,
             databaseBuilder: databaseBuilder);
 
-  Future<List<User>> getSfriends() async {
+  Future<List<User>> getFollowers() async {
     return (await getSUsers())
         .where((User u) => u.membership == Membership.join)
         .toList();
@@ -60,23 +75,27 @@ class SClient extends Client {
   SMatrixRoom userRoom;
   bool get userRoomCreated => userRoom != null;
 
+  Timer timerCallbackRoomUpdate;
+  Timer timerCallbackEventUpdate;
   Future<void> initSMatrix() async {
     // initialisation
 
     await loadSRooms();
-    await sendInvitesToFriends();
+    await autoFollowFollowers(); // TODO : Let's see if we keep this in the future
     await loadNewTimeline();
+    notifications.loadNotifications(this);
 
-    onEventUpdate ??= this.onEvent.stream.listen((EventUpdate eUp) async {
-      print(eUp.type);
-      if (eUp.eventType == "m.room.message") {
-        await loadNewTimeline();
-      }
-    });
-
+    // for smatrixrooms
     onRoomUpdateSub ??= this.onRoomUpdate.stream.listen((RoomUpdate rUp) async {
-      await loadSRooms();
-      await loadNewTimeline();
+      if (srooms.containsKey(rUp.id)) {
+        // we use a timer to prevent calling
+        timerCallbackRoomUpdate?.cancel();
+        timerCallbackRoomUpdate =
+            new Timer(const Duration(milliseconds: 500), () async {
+          log.info("MinesTrix : callback new message");
+          await loadNewTimeline(); // new message, we only need to rebuild timeline
+        });
+      }
     });
   }
 
@@ -84,29 +103,63 @@ class SClient extends Client {
     await loadSTimeline();
     sortTimeline();
 
-    onTimelineUpdate.add("up");
-    //await onTimelineUpdate.done;
+    notifications.loadNotifications(this);
 
+    // On first sync we fetch all the event history
     if (_firstSync) {
-      Duration duration = Duration(seconds: 2); // let the app start
-      _timer = Timer(duration, () async {
-        print("Timer, sync threads");
+      try {
+        int n = srooms.values.length;
+        int counter = 0;
         for (SMatrixRoom sr in srooms.values) {
           await sr.timeline.requestHistory();
+
+          log.info("First sync progress : " + (counter / n * 100).toString());
+          counter++;
         }
-      });
+      } catch (e) {
+        log.severe("Initial sync : failed to get history", e);
+      }
       _firstSync = false;
     }
+    onTimelineUpdate.add("up");
   }
 
-  void setupSRoom(SMatrixRoom sroom) async {
-    if (sroom.room.name.startsWith(SMatrixUserRoomPrefix)) {
-      print("setup room");
-      String roomName = sroom.room.name.replaceFirst("smatrix_", "");
-      await sroom.room.setName(roomName + " timeline");
+  Future<bool> trySettingRoomState(Room room) async {
+    try {
+      Event e = room.getState("org.matrix.msc1840");
+
+      if (e == null || e.content["type"] != "fr.henri2h.minestrix") {
+        Map<String, dynamic> content = new Map<String, dynamic>();
+        content["type"] = "fr.henri2h.minestrix";
+        await this.setRoomStateWithKey(room.id, "org.matrix.msc1840", content);
+        return true;
+      }
+    } catch (e) {
+      log.severe("could not set user thread room as minestrix room", e);
+      return false;
+    }
+    return true; // no update needed
+  }
+
+// setup the user room
+  Future<bool> setupSRoom(SMatrixRoom sroom) async {
+    try {
+      // migration
+      // TODO : remove this if statement in a future release
+      if (sroom.room.name.startsWith(SMatrixUserRoomPrefix)) {
+        log.warning("Update sroom");
+        String roomName = sroom.room.name.replaceFirst("smatrix_", "");
+        await sroom.room.setName(roomName + " timeline");
+      }
+      await trySettingRoomState(sroom.room);
+      return true;
+    } catch (e) {
+      log.severe("Error setup sroom", e);
+      return false;
     }
   }
 
+  bool sroomsLoaded = false;
   Future<void> loadSRooms() async {
     // userRoom = null; sometimes an update miss the user room... in order to prevent indesired refresh we suppose that the room won't be removed.
     // if the user room is removed, the user should restart the app
@@ -118,25 +171,28 @@ class SClient extends Client {
 
     for (var i = 0; i < rooms.length; i++) {
       Room r = rooms[i];
+
       if (r.membership == Membership.invite) {
-        print("Friendship requests sent : " + r.name);
+        log.info("Friendship requests sent : " + r.name);
       }
+
       SMatrixRoom rs = SMatrixRoom();
       if (await rs.init(r, this)) {
-        print(rs.room.name);
         // if class is correctly initialisated, we can add it
         // if we are here, it means that we have a valid smatrix room
+
         if (r.membership == Membership.join) {
           rs.timeline = await rs.room.getTimeline();
           srooms[rs.room.id] = rs;
 
           // by default
           if (rs.room.pushRuleState == PushRuleState.notify)
-            await rs.room.setPushRuleState(PushRuleState.mentions_only);
+            await rs.room.setPushRuleState(PushRuleState.mentionsOnly);
           if (!rs.room.tags.containsKey("m.lowpriority")) {
             await rs.room.addTag("m.lowpriority");
           }
 
+          // check if this room is a user thread
           if (rs.roomType == SRoomType.UserRoom) {
             userIdToRoomId[rs.user.id] = rs.room.id;
 
@@ -156,40 +212,49 @@ class SClient extends Client {
       }
     }
 
-    if (userRoom == null) print("❌ User room not found");
+    onSRoomsUpdate.add("update");
+    sroomsLoaded = true;
+
+    if (userRoom == null) log.severe("❌ User room not found");
   }
 
-  Future createSMatrixRoom() async {
-    print("Create smatrix room");
+  Future<SMatrixRoom> createSMatrixRoom(String name, String desc) async {
     String roomID = await createRoom(
-        name: "@" + userID + " timeline",
-        topic: "Mines'Trix room name",
-        visibility: Visibility.private);
-
+        name: name, topic: desc, visibility: Visibility.private);
     SMatrixRoom sroom = SMatrixRoom();
+
     Room r = getRoomById(roomID);
+    await trySettingRoomState(r);
+
     bool result = await sroom.init(r, this);
 
-    /*
-    Map<String, dynamic> content = Map<String, dynamic>();
-    content["type"] = "fr.henri2h.smatirx";
-    String res = await r.sendEvent(content,
-        type: "org.matrix.msc1840"); // define room type, MSC 1840
-    print(res); 
-    // save user room*/ // NOT working...
-    if (result)
-      userRoom = sroom;
-    else
-      print("could not creat room ....");
+    await setupSRoom(sroom); // add the room type
+
+    if (result) {
+      // launch sync
+      await loadSRooms();
+      await loadNewTimeline();
+
+      return sroom;
+    } else
+      return null;
+  }
+
+  Future createSMatrixUserProfile() async {
+    log.info("Create smatrix room");
+    String name = userID + " timeline";
+    SMatrixRoom sroom = await createSMatrixRoom(name, "Mines'Trix room name");
+
+    if (sroom != null) userRoom = sroom;
   }
 
   Iterable<Event> getSRoomFilteredEvents(Timeline t) {
     List<Event> filteredEvents = t.events
         .where((e) =>
             !{
-              RelationshipTypes.Edit,
-              RelationshipTypes.Reaction,
-              RelationshipTypes.Reply
+              RelationshipTypes.edit,
+              RelationshipTypes.reaction,
+              RelationshipTypes.reply
             }.contains(e.relationshipType) &&
             {EventTypes.Message, EventTypes.Encrypted}.contains(e.type) &&
             !e.redacted)
@@ -216,18 +281,18 @@ class SClient extends Client {
       return b.originServerTs.compareTo(a.originServerTs);
     });
 
-    print("stimeline length : " + stimeline.length.toString());
+    log.info("stimeline length : " + stimeline.length.toString());
   }
 
   /* this function iterate over all accepted friends invitations and ensure that they are in the user room
   then it accepts all friends invitations from members of the user room
     */
-  Future<void> sendInvitesToFriends() async {
-    List<User> friends = await getSfriends();
+  Future<void> autoFollowFollowers() async {
+    List<User> followers = await getFollowers();
     List<SMatrixRoom> sr = sInvites.values.toList();
     for (SMatrixRoom r in sr) {
       // check if the user is already in the list and accept invitation
-      bool exists = (friends.firstWhere((element) => r.user.id == element.id,
+      bool exists = (followers.firstWhere((element) => r.user.id == element.id,
               orElse: () => null) !=
           null);
       if (exists) {
