@@ -1,6 +1,8 @@
 import 'package:flutter/material.dart';
+import 'package:logging/logging.dart';
 import 'package:minestrix/app.dart';
 
 void main() {
-  runApp(Minetrix());
+  Logger.root.level = Level.ALL;
+  runApp(Minestrix());
 }
