@@ -21,7 +21,7 @@ class FriendsVue extends StatelessWidget {
 
     return ListView(
       children: [
-        H1Title("Users"),
+        H1Title("Following"),
         Padding(
           padding: const EdgeInsets.all(8.0),
           child: TypeAheadField(
@@ -30,12 +30,16 @@ class FriendsVue extends StatelessWidget {
                 autofocus: false,
                 decoration: InputDecoration(border: OutlineInputBorder())),
             suggestionsCallback: (pattern) async {
-              UserSearchResult ur = await sclient.searchUser(pattern);
-              List<User> sFriends = await sclient.getSfriends();
+              UserSearchResult ur = await sclient.searchUserDirectory(pattern);
+
+              List<User> following = List<User>.empty();
+              await sclient.following.forEach((key, SMatrixRoom sroom) {
+                following.add(sroom.user);
+              });
 
               return ur.results
                   .where((element) =>
-                      sFriends.firstWhere(
+                      following.firstWhere(
                           (friend) => friend.id == element.userId,
                           orElse: () => null) ==
                       null)
@@ -53,7 +57,6 @@ class FriendsVue extends StatelessWidget {
             },
             onSuggestionSelected: (suggestion) async {
               Profile p = suggestion;
-              print(p.userId);
               await sclient.addFriend(p.userId);
             },
           ),
