@@ -4,13 +4,17 @@
 
 #include "generated_plugin_registrant.h"
 
-#include <file_chooser/file_chooser_plugin.h>
+#include <file_selector_linux/file_selector_plugin.h>
+#include <flutter_secure_storage/flutter_secure_storage_plugin.h>
 #include <url_launcher_linux/url_launcher_plugin.h>
 
 void fl_register_plugins(FlPluginRegistry* registry) {
-  g_autoptr(FlPluginRegistrar) file_chooser_registrar =
-      fl_plugin_registry_get_registrar_for_plugin(registry, "FileChooserPlugin");
-  file_chooser_plugin_register_with_registrar(file_chooser_registrar);
+  g_autoptr(FlPluginRegistrar) file_selector_linux_registrar =
+      fl_plugin_registry_get_registrar_for_plugin(registry, "FileSelectorPlugin");
+  file_selector_plugin_register_with_registrar(file_selector_linux_registrar);
+  g_autoptr(FlPluginRegistrar) flutter_secure_storage_registrar =
+      fl_plugin_registry_get_registrar_for_plugin(registry, "FlutterSecureStoragePlugin");
+  flutter_secure_storage_plugin_register_with_registrar(flutter_secure_storage_registrar);
   g_autoptr(FlPluginRegistrar) url_launcher_linux_registrar =
       fl_plugin_registry_get_registrar_for_plugin(registry, "UrlLauncherPlugin");
   url_launcher_plugin_register_with_registrar(url_launcher_linux_registrar);
